+++ dependency-reduced-pom.xml
@@ -3,7 +3,7 @@
   <modelVersion>4.0.0</modelVersion>
   <groupId>Seisan</groupId>
   <artifactId>Seisan_Plugin_For_Ninkai</artifactId>
-  <version>3.0.3</version>
+  <version>3.0.4</version>
   <build>
     <defaultGoal>clean package</defaultGoal>
     <resources>
